Update Managed X10 native bindings for FT to work with Rail[Complex] instead of Rail[Double].
git-svn-id: svn://svn.code.sf.net/p/x10/code/benchmarks/trunk@27694 a326200e-df2c-42e0-81fa-597e909af41d

+++ PERCS/FT/FTNatives.cc
@@ -9,8 +9,8 @@
  * Method:    create_plan
  * Signature: (III)J
  */
-JNIEXPORT jlong JNICALL Java_FTNatives_create_1plan(JNIEnv *env, jclass klass, jint SQRTN, jint direction, jint flags) {
-  return create_plan(SQRTN, direction, flags);
+JNIEXPORT jlong JNICALL Java_FTNatives_create_1plan(JNIEnv *env, jclass klass, jlong SQRTN, jint direction, jint flags) {
+    return create_plan((jint)SQRTN, direction, flags);
 }
 
 /*
@@ -18,11 +18,11 @@ JNIEXPORT jlong JNICALL Java_FTNatives_create_1plan(JNIEnv *env, jclass klass, j
  * Method:    execute_plan
  * Signature: (J[D[DIII)V
  */
-JNIEXPORT void JNICALL Java_FTNatives_execute_1plan(JNIEnv *env, jclass klass, jlong plan, jdoubleArray Ajava, jdoubleArray Bjava, jint SQRTN, jint i0, jint i1) {
+JNIEXPORT void JNICALL Java_FTNatives_execute_1plan(JNIEnv *env, jclass klass, jlong plan, jdoubleArray Ajava, jdoubleArray Bjava, jlong SQRTN, jlong i0, jlong i1) {
   jdouble *A = env->GetDoubleArrayElements(Ajava, (jboolean*)NULL);
   jdouble *B = env->GetDoubleArrayElements(Bjava, (jboolean*)NULL);
 
-  execute_plan(plan, A, B, SQRTN, i0, i1);
+  execute_plan(plan, A, B, (jint)SQRTN, (jint)i0, (jint)i1);
 
   env->ReleaseDoubleArrayElements(Ajava, A, 0);
   env->ReleaseDoubleArrayElements(Bjava, B, 0);
